chore: move buildInfo
