+++ cmd/celestia/util.go
@@ -4,7 +4,6 @@ import (
 	"github.com/spf13/cobra"
 
 	cmdnode "github.com/celestiaorg/celestia-node/cmd"
-	"github.com/celestiaorg/celestia-node/cmd/version"
 	"github.com/celestiaorg/celestia-node/nodebuilder/core"
 	"github.com/celestiaorg/celestia-node/nodebuilder/gateway"
 	"github.com/celestiaorg/celestia-node/nodebuilder/header"
@@ -27,13 +26,6 @@ func persistentPreRunEnv(cmd *cobra.Command, nodeType node.Type, _ []string) err
 		return err
 	}
 	ctx = cmdnode.WithNetwork(ctx, parsedNetwork)
-	buildInfo := version.GetBuildInfo()
-	ctx = cmdnode.WithNodeBuildInfo(ctx, &node.BuildInfo{
-		LastCommit:      buildInfo.LastCommit,
-		SemanticVersion: buildInfo.SemanticVersion,
-		SystemVersion:   buildInfo.SystemVersion,
-		GolangVersion:   buildInfo.GolangVersion,
-	})
 
 	// loads existing config into the environment
 	ctx, err = cmdnode.ParseNodeFlags(ctx, cmd, cmdnode.Network(ctx))
