+++ cmd/env.go
@@ -38,11 +38,6 @@ func NodeConfig(ctx context.Context) nodebuilder.Config {
 	return cfg
 }
 
-// NodeInfo reads the node build inforamtion from the context.
-func NodeInfo(ctx context.Context) node.BuildInfo {
-	return ctx.Value(buildInfo{}).(node.BuildInfo)
-}
-
 // WithNodeType sets the node type in the given context.
 func WithNodeType(ctx context.Context, tp node.Type) context.Context {
 	return context.WithValue(ctx, nodeTypeKey{}, tp)
@@ -78,16 +73,10 @@ func WithNodeConfig(ctx context.Context, config *nodebuilder.Config) context.Con
 	return context.WithValue(ctx, configKey{}, *config)
 }
 
-// WithNodeBuildInfo sets the node config build information.
-func WithNodeBuildInfo(ctx context.Context, info *node.BuildInfo) context.Context {
-	return context.WithValue(ctx, node.BuildInfo{}, *info)
-}
-
 type (
 	optionsKey   struct{}
 	configKey    struct{}
 	storePathKey struct{}
 	nodeTypeKey  struct{}
 	networkKey   struct{}
-	buildInfo    struct{}
 )
