+++ nodebuilder/node/buildInfo.go
@@ -1,9 +1,35 @@
 package node
 
-// BuildInfo stores all necessary information for the current build.
+import (
+	"fmt"
+	"runtime"
+)
+
+var (
+	buildTime       string
+	lastCommit      string
+	semanticVersion string
+
+	systemVersion = fmt.Sprintf("%s/%s", runtime.GOARCH, runtime.GOOS)
+	golangVersion = runtime.Version()
+)
+
+// BuildInfo represents all necessary information about current build.
 type BuildInfo struct {
+	BuildTime       string
 	LastCommit      string
 	SemanticVersion string
 	SystemVersion   string
 	GolangVersion   string
+}
+
+// GetBuildInfo returns information about current build.
+func GetBuildInfo() *BuildInfo {
+	return &BuildInfo{
+		buildTime,
+		lastCommit,
+		semanticVersion,
+		systemVersion,
+		golangVersion,
+	}
 }
